+++ be/src/main/java/airdnb/be/domain/member/MemberRepository.java
@@ -1,0 +1,10 @@
+package airdnb.be.domain.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    boolean existsByEmail(String email);
+}
