--- be/src/main/java/airdnb/be/member/MemberRepository.java
@@ -1,8 +1,0 @@
-package airdnb.be.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-}
