--- be/src/main/java/airdnb/be/member/controller/MemberController.java
@@ -1,23 +1,0 @@
-package airdnb.be.member.controller;
-
-import airdnb.be.member.MemberService;
-import airdnb.be.member.controller.dto.MemberDto;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/member")
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @PostMapping
-    public void add(@Valid @RequestBody MemberDto dto) {
-        memberService.add(dto);
-    }
-}
