+++ be/src/main/java/airdnb/be/web/controller/member/MemberController.java
@@ -1,0 +1,34 @@
+package airdnb.be.web.controller.member;
+
+import airdnb.be.member.service.MemberService;
+import airdnb.be.web.controller.member.request.EmailRequest;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/member")
+public class MemberController {
+
+    private final MemberService memberService;
+
+    /**
+     * @return HttpStatus.OK 는 로그인 창 HttpStatus.NO_CONTENT 는 회원가입
+     */
+    @GetMapping("/exist")
+    public ResponseEntity<Void> existsMemberByEmail(@RequestBody @Valid EmailRequest request) {
+        boolean existsMember = memberService.existsMemberByEmail(request.email());
+        if (existsMember) {
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+}
